+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ### Added
 - Added `ProjectFilterOptions` as a possible input for the `myProjects` resolver
+- Added `PlanFeedback` and `PlanFeedbackComments` models [#243]
+- Added `projectCollaboratorCommentsAdded` to `emailService` so that we can email `project collaborators` when new comments added [#243]
+- Added comment mutations to `answers` resolver and created `feedback` resolver [#243]
 - data migration to change collation to `utf8mb4` on all tables
 - added `publishedSection` resolver to `src/resolvers/versionedSection.ts`
 - added `publishedQuestion` resolver to `src/resolvers/versionedQuestion.ts`